Changes to be committed: 	modified:   server/controller/admin/BrandsController.js 	modified:   server/controller/admin/DeviceController.js 	modified:   server/models/Brands.js 	modified:   server/models/Devices.js 	new file:   server/routes/admin/deviceRouter.js 	modified:   server/routes/admin/metersRouter.js

+++ server/models/Devices.js
@@ -19,45 +19,70 @@ const Devices = sequelize.define("devices", {
 Devices.createDevice = async ({ name }) => {
     try {
         const device = await Devices.create({ name });
-        return device;
+        if (!device) {
+            throw new Error("Устройство не добавлено");
+        }
+        return { device };
     } catch (e) {
         console.error(e);
+        return { error: e.message, status: 404 };
     }
 };
 
 Devices.getOneDeviceById = async (id) => {
     try {
         const device = await Devices.findByPk(id);
+        if (!device) {
+            throw new Error("Устройство не найдено");
+        }
         return device;
     } catch (e) {
         console.error(e);
+        return { error: e.message, status: 404 };
     }
 };
 
 Devices.getAllDevices = async () => {
     try {
         const devices = await Devices.findAll();
-        return devices;
+        if (devices.length === 0) {
+            throw new Error("Устройства не найдены");
+        }
+        return { devices };
     } catch (e) {
         console.error(e);
+        return { error: e.message, status: 404 };
     }
 };
 
 Devices.deleteDevice = async (id) => {
     try {
-        const device = await Devices.destroy({ where: { id } });
-        return device;
+        const device = await Devices.findByPk(id);
+        if (!device) {
+            throw new Error("Устройство не найдено");
+        }
+        await Devices.destroy({ where: { id } });
+        return { device };
     } catch (e) {
         console.error(e);
+        return { error: e.message, status: 404 };
     }
 };
 
 Devices.updateDevice = async ({ id, name }) => {
     try {
-        const device = await Devices.update({ name }, { where: { id } });
-        return device;
+        const device = await Devices.findByPk(id);
+        if (!device) {
+            throw new Error("Устройство не найдено");
+        }
+        const newDevice = await Devices.update({ name }, { where: { id } });
+        if (!newDevice) {
+            throw new Error("Устройство не обновлено");
+        }
+        return { device };
     } catch (e) {
         console.error(e);
+        return { error: e.message, status: 404 };
     }
 };
